Add the keycloak deployment as an implicit argument
When you create a new keycloak deployment, it populates the OIDC
configuration by calling the keycloak server. Because we were creating
one inside the token method, the server was being called every time we
checked a token which is a waste of time because that information almost
never changes.

I've added individual implicits to the methods rather than to the whole
class mainly because it's easier to test with the test libraries that
we're using but also, it is possible to use the same global keycloak
utils instance to get tokens from different servers. It's unlikely but
it still makes this a bit more flexible.